CPR-1148 updated test assertions
