+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/message/listeners/probation/ProbationAddressCreatedEventListenerIntTest.kt
@@ -2,10 +2,6 @@ package uk.gov.justice.digital.hmpps.personrecord.message.listeners.probation
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.ProbationAddress
-import uk.gov.justice.digital.hmpps.personrecord.model.types.AddressStatusCode
-import uk.gov.justice.digital.hmpps.personrecord.model.types.AddressUsageCode
-import uk.gov.justice.digital.hmpps.personrecord.model.types.ContactType
 import uk.gov.justice.digital.hmpps.personrecord.service.type.OFFENDER_ADDRESS_CREATED
 import uk.gov.justice.digital.hmpps.personrecord.test.randomCrn
 
@@ -23,7 +19,8 @@ class ProbationAddressCreatedEventListenerIntTest : ProbationEventListenerTestBa
 
     publishProbationAddressEvent(cprPerson.crn, probationAddress.deliusAddressId, OFFENDER_ADDRESS_CREATED)
 
-    assertAddressSaved(cprPerson.crn!!, probationAddress)
+    val actualPersonEntity = awaitNotNull { personRepository.findByCrn(cprPerson.crn!!) }
+    assertAddress(actualPersonEntity, probationAddress)
   }
 
   @Test
@@ -54,35 +51,5 @@ class ProbationAddressCreatedEventListenerIntTest : ProbationEventListenerTestBa
     expectNoMessagesOn(probationEventsQueue)
     expectOneMessageOnDlq(probationEventsQueue)
     assertThat(personRepository.findByCrn(cprPerson.crn!!)!!.addresses.size).isEqualTo(0)
-  }
-
-  private fun assertAddressSaved(crn: String, probationAddress: ProbationAddress) {
-    awaitAssert {
-      val actualPersonEntity = personRepository.findByCrn(crn)!!
-      assertThat(actualPersonEntity.addresses.size).isEqualTo(1)
-      val actualAddressEntity = actualPersonEntity.addresses.first()
-      assertThat(actualAddressEntity.updateId).isNotNull()
-      assertThat(actualAddressEntity.updateId!!.toString()).isNotBlank
-      assertThat(actualAddressEntity.noFixedAbode).isEqualTo(probationAddress.noFixedAbode)
-      assertThat(actualAddressEntity.startDate).isEqualTo(probationAddress.startDate)
-      assertThat(actualAddressEntity.endDate).isEqualTo(probationAddress.endDate)
-      assertThat(actualAddressEntity.postcode).isEqualTo(probationAddress.postcode)
-      assertThat(actualAddressEntity.fullAddress).isEqualTo(probationAddress.fullAddress)
-      assertThat(actualAddressEntity.buildingName).isEqualTo(probationAddress.buildingName)
-      assertThat(actualAddressEntity.postTown).isEqualTo(probationAddress.townCity)
-      assertThat(actualAddressEntity.county).isEqualTo(probationAddress.county)
-      assertThat(actualAddressEntity.uprn).isEqualTo(probationAddress.uprn)
-      assertThat(actualAddressEntity.deliusAddressId).isEqualTo(probationAddress.deliusAddressId)
-      assertThat(actualAddressEntity.isVerified).isEqualTo(probationAddress.isVerified)
-      assertThat(actualAddressEntity.usages.first().usageCode).isEqualTo(AddressUsageCode.from(probationAddress.usage!!.code))
-      assertThat(actualAddressEntity.statusCode).isEqualTo(AddressStatusCode.valueOf(probationAddress.status?.code!!))
-
-      assertThat(actualAddressEntity.buildingNumber).isEqualTo(probationAddress.addressNumber)
-      assertThat(actualAddressEntity.thoroughfareName).isEqualTo(probationAddress.streetName)
-      assertThat(actualAddressEntity.dependentLocality).isEqualTo(probationAddress.district)
-      assertThat(actualAddressEntity.comment).isEqualTo(probationAddress.notes)
-      assertThat(actualAddressEntity.contacts.first().contactType).isEqualTo(ContactType.HOME)
-      assertThat(actualAddressEntity.contacts.first().contactValue).isEqualTo(probationAddress.telephoneNumber)
-    }
   }
 }
